requirements.txt
gunicorn
django-heroku==0.3.1
pandas==0.23.4
google-cloud-bigquery==1.7.0
django-cors-headers==2.4.0
simplejson==3.16.0
